tools/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic[email]==2.10.3
pydantic-settings==2.6.1
PyJWT[crypto]==2.10.1
passlib[bcrypt]==1.7.4
boto3==1.35.86
python-multipart==0.0.22
httpx==0.28.1
pyotp==2.9.0
